Shutdown cleanly
Fixes #46

If there are still entries to be read from the journal when fluend
shuts down we will continue emitting records even after the router has
been removed.

This change stops reading after the shutdown hook gets called.

It also has the nice side effect that we stop updating pos storage,
so we shouldn't miss any entries when / if we start up again.