Use latest systemd-journal
There is known bug that systemd-journal 1.4.2 or older version
doesn't work with custom memory allocator such as jemalloc because of
inappropriate allocated memory handling.
As a result, it causes a SEGV. This bug was frequently reported from
fluent-docker-image and fluentd-kubernetes-daemonset images users. [1]

[1] https://github.com/fluent/fluentd-docker-image/issues/378

Recently, this bug was fixed [2] and released as systemd-journal 2.0.0.
Now we should switch to it.

[2] https://github.com/ledbettj/systemd-journal/pull/97

NOTE: systemd-journal requires Ruby 3.0.0 or later, so need to bump
base image version which provides Ruby 3.x - so use ubuntu:jammy for
testing.

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>